corrected namespace of automation tests
